+++ src/services/SoupService.ts
@@ -1,33 +1,32 @@
-
 import Ingredient from "../models/Ingredient";
-import Recipe from "../models/Recipe";
+import Soup from "../models/Soup";
 import RecipeMatrixItem from "../models/RecipeMatrixItem";
-import ingredientsJson from '../ingredients.json';
-import recipeJson from "../recipe_matrix.json";
+import ingredientsJson from "../data/ingredients.json";
+import recipeJson from "../data/soupMatrix.json";
 
 const ingredients: Ingredient[] = ingredientsJson;
+
 const recipeMatrix: RecipeMatrixItem[][] = recipeJson;
 
-function getRankedRecipes(selectedIngredients: Ingredient[]) : Recipe[] {
+function getRankedSoups(selectedIngredients: Ingredient[]): Soup[] {
   const pairs: [Ingredient, Ingredient][] = [];
-  
+
   selectedIngredients.forEach((ingredient, index) => {
     const rightIngredients = selectedIngredients.slice(index);
-    
+
     rightIngredients.forEach((rightIngredient) => {
       if (ingredient.index >= rightIngredient.index)
         pairs.push([ingredient, rightIngredient]);
-      else
-        pairs.push([rightIngredient, ingredient])
+      else pairs.push([rightIngredient, ingredient]);
     });
   });
 
-  const recipes: Recipe[] = [];
+  const recipes: Soup[] = [];
 
   pairs.forEach((pair) => {
     recipes.push({
       ingredients: [pair[0], pair[1]],
-      ...recipeMatrix[pair[0].index][pair[1].index]
+      ...recipeMatrix[pair[0].index][pair[1].index],
     });
   });
 
@@ -36,4 +35,4 @@ function getRankedRecipes(selectedIngredients: Ingredient[]) : Recipe[] {
   return recipes;
 }
 
-export {ingredients, recipeMatrix, getRankedRecipes};
+export { ingredients, recipeMatrix, getRankedSoups };
